requirements.txt
Authlib==1.3.0
black==24.2.0
blinker==1.7.0
cachelib==0.9.0
certifi==2024.2.2
cffi==1.16.0
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.5
distlib==0.3.8
dnspython==2.6.1
filelock==3.13.1
Flask==3.0.2
Flask-Caching==2.1.0
Flask-Login==0.6.3
Flask-Session==0.6.0
Flask-WTF==1.2.1
identify==2.5.35
idna==3.6
isort==5.13.2
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
mypy-extensions==1.0.0
nodeenv==1.8.0
packaging==23.2
pandas==2.2.1
pathspec==0.12.1
platformdirs==4.2.0
pre-commit==3.6.2
pycparser==2.21
pymongo==4.6.2
python-decouple==3.8
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
urllib3==2.2.1
virtualenv==20.25.1
Werkzeug==3.0.1
WTForms==3.1.2
